% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexconvol.R
\name{hexconvol}
\alias{hexconvol}
\title{Compute convolution over neighbours of elements of a hexmatrix.}
\usage{
hexconvol(m, probs = 1)
}
\arguments{
\item{m}{a numeric hexmatrix}

\item{probs}{a vector of length 7 with six weights corresponding to the neighbours
clock-wisely from top-left. The 7th value corresponds to the weight of a value
at the "current" position. Precisely, the weights are understood as follows:
\enumerate{
   \item `probs[1]` for top-left,
   \item `probs[2]` for top-right,
   \item `probs[3]` for right,
   \item `probs[4]` for bottom-right,
   \item `probs[5]` for bottom-left,
   \item `probs[6]` for left neighbour's weight of the "current" position,
   \item `probs[7]` for the "current" value's weight.
}}
}
\value{
a hexmatrix of dimensions equal to `m` with values computed as a convolution
    over the neighbours of `m`'s values
}
\description{
The result of this function is a hexmatrix of dimensions equal to `m` with values
computed as a weighted sum of value at the given position and in the neighbourhood
of that position. The weights are given by the `probs` argument.
}
