% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R, R/point.R
\name{index}
\alias{index}
\alias{index.matrix}
\alias{index.array}
\alias{point.matrix}
\alias{point.array}
\title{Get index an element of a matrix or array given by the point coordinates,
i.e., row, column, (and layer, for arrays) number.}
\usage{
index(m, point, ...)

\method{index}{matrix}(m, point, ...)

\method{index}{array}(m, point, ...)

\method{point}{matrix}(m, index, ...)

\method{point}{array}(m, index, ...)
}
\arguments{
\item{m}{A matrix or an array whose element index has to be computed}

\item{point}{The coordinates of the element. For matrix, it should be a
vector of two values (row, column). For array, three values have to be
provided (row, column, layer).}
}
\value{
An integer value corresponding to the index of the element determined
in \code{point}, if \code{m} was treated as a single-dimensional vector.
}
\description{
This function is an inverse function to \code{\link[=point]{point()}}.
}
\seealso{
\code{\link[=point]{point()}}
}
