% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like.R
\name{like}
\alias{like}
\alias{like.matrix}
\alias{like.array}
\title{Create a copy of given data and possibly change some attributes of the new
instance.}
\usage{
like(...)

\method{like}{matrix}(
  m,
  data = .empty,
  nrow = .empty,
  ncol = .empty,
  byrow = FALSE,
  dimnames = .empty
)

\method{like}{array}(a, data = .empty, dim = .empty, dimnames = .empty)
}
\arguments{
\item{m}{A matrix to be copied}

\item{data}{New data to be stored in the new copy}

\item{nrow}{The new number of rows of the copied matrix}

\item{ncol}{The new number of columns of the copied matrix}

\item{byrow}{Whehter the data of the copied matrix has to be assigned by row
first}

\item{dimnames}{THe dimension names of the new copy}

\item{a}{An array to be copied}

\item{dim}{The dimensions of the new copied array}
}
\description{
Create a copy of given data and possibly change some attributes of the new
instance.
}
\examples{
m <- matrix(1:6, nrow=3)

# create a new matrix of the same dimensions as m, but different data
m2 <- like(m, data=0)
}
