% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize the values of the array by given margin}
\usage{
normalize(x, MARGIN = seq_len(length(dim(x) - 1)), SUM = sum, DIV = `/`)
}
\arguments{
\item{x}{An array to be normalized}

\item{MARGIN}{A margin, over which to perform the normalization.}

\item{SUM}{The summationfunction}

\item{DIV}{The division function}
}
\value{
An array of dimensions equal to `x` with normalized values
}
\description{
By normalization we understand dividing each value of the cell by a sum
computed along the given `MARGIN`. The summation and division may be replaced
with custom functions `SUM` and `DIV`, respectively.
}
\examples{
m <- matrix(1:12, nrow=4)
r <- normalize(m, 1)   # normalize over the rows
rowSums(r)             # returns a vector of 1s
s <- normalize(m, 2)   # normalize over the columns
colSums(s)             # returns a vector of 1s
}
