% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer2.R
\name{outer2}
\alias{outer2}
\title{Outer product of multiple arrays}
\usage{
outer2(f, ...)
}
\arguments{
\item{f}{The function to be used on the outer products}

\item{...}{Vectors, matrices or arrays whose values will be used as arguments
to the function `f`. The elements of the first argument will be used as
the first argument to `f`, elements of the second argument will be used as
the second argument to `f` etc.}
}
\value{
An array that is an outer product of the input arguments.
}
\description{
This function works similarly as `base::outer()` except that it allows the
multiplication of more arguments than just two.
}
\examples{
outer2(prod, 1:4)           # returns 1:4
outer2(prod, 1:4, 1:3)      # the same as 1:4 \%o\% 1:3
outer2(prod, 1:4, 1:3, 1:2) # the same as (1:4 \%o\% 1:3) \%o\% 1:2

# to see how f is called with arguments:
outer2(paste, 1:4, 1:3, 1:2)
}
