% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{Get coordinates (i.e., row, column, (layer)) of an element of a matrix or
array given by the index.}
\usage{
point(m, index, ...)
}
\arguments{
\item{m}{A matrix or an array whose element coordinates have to be computed}

\item{index}{The index of the element.}
}
\value{
For matrix, it returns a vector of two values (row, column). For
    array, a vector of three values is returned (row, column, layer).
}
\description{
This function is an inverse function to [index()].
}
\seealso{
[index()]
}
