% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Find a region of values in a hexmatrix}
\usage{
region(m, i)
}
\arguments{
\item{m}{A hexmatrix of values}

\item{i}{A index of a cell to start from}
}
\value{
A vector of indices of cells that are neighbours of `i` and have the
  same value in `m` as `m[i]`, or cells recursively neighbouring with such
  neighbours.
}
\description{
For a given hexmatrix and a given index of a cell, find all indices of all
neighbours that have the same value.
}
