% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Rescale the given hexmatrix by factor of 0.5. The resulting hexmatrix would
have half of rows and half of columns of the original matrix. The values in
cells will be summed appropriately.}
\usage{
shrink(m)
}
\arguments{
\item{m}{A hexmatrix to shrink.}
}
\value{
A rescaled hexmatrix.
}
\description{
Rescale the given hexmatrix by factor of 0.5. The resulting hexmatrix would
have half of rows and half of columns of the original matrix. The values in
cells will be summed appropriately.
}
