% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/via.R
\name{via}
\alias{via}
\title{Return a logical hexarray of paths that lead through the given intermediate
node}
\usage{
via(paths, intermed)
}
\description{
Return a logical hexarray of paths that lead through the given intermediate
node
}
