% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexconvol.R
\name{hexconvol}
\alias{hexconvol}
\title{Compute convolution over neighbours of elements of a hexmatrix.}
\usage{
hexconvol(m, probs = 1)
}
\arguments{
\item{m}{a numeric hexmatrix}

\item{probs}{a vector of length 7 with six weights corresponding to the neighbours
clock-wisely from top-left. The 7th value corresponds to the weight of a value
at the "current" position. Precisely, the weights are understood as follows:
\enumerate{
\item \code{probs[1]} for top-left,
\item \code{probs[2]} for top-right,
\item \code{probs[3]} for right,
\item \code{probs[4]} for bottom-right,
\item \code{probs[5]} for bottom-left,
\item \code{probs[6]} for left neighbour's weight of the "current" position,
\item \code{probs[7]} for the "current" value's weight.
}}
}
\value{
a hexmatrix of dimensions equal to \code{m} with values computed as a convolution
over the neighbours of \code{m}'s values
}
\description{
The result of this function is a hexmatrix of dimensions equal to \code{m} with values
computed as a weighted sum of value at the given position and in the neighbourhood
of that position. The weights are given by the \code{probs} argument.
}
