% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexconvol.R
\name{hexconvol}
\alias{hexconvol}
\title{Compute convolution over the neighbours.}
\usage{
hexconvol(m, probs = 1)
}
\arguments{
\item{m}{a hex-matrix}

\item{probs}{a vector of length 7 with probabilities given to the neighbours
clock-wisely from top-left with "this" coordinate's probability at 7th
position}
}
\description{
Compute convolution over the neighbours.
}
