% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiking.R
\name{hiking}
\alias{hiking}
\title{Tobler's hiking function}
\usage{
hiking(gradient)
}
\arguments{
\item{gradient}{a fraction of elevation difference over the horizontal distance
(also a tangens of the slope angle)}
}
\value{
a walking velocity in km/h
}
\description{
A function that determines the hiking speed of a typical healthy adult person while
taking into account the slope angle.
}
\examples{
hiking(-0.05)   # a maximum of 6 km/h corresponding to the downhill angle of 2.86 degrees
hiking(0)       # speed on a path with flat surface
}
\references{
Tobler, Waldo (February 1993). Three presentations on geographical
    analysis and modeling: Non-isotropic geographic modeling speculations on the
    geometry of geography global spatial analysis. Technical Report. National
    center for geographic information and analysis. 93 (1).
}
