% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{neighbours}
\alias{neighbours}
\title{Returns a list of matrices that represent values of neighbours
in the clock-wise order starting from top-left neighbour}
\usage{
neighbours(m, self = FALSE)
}
\description{
Returns a list of matrices that represent values of neighbours
in the clock-wise order starting from top-left neighbour
}
