% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichDir.R
\name{whichDir}
\alias{whichDir}
\title{Return the number of direction from the current to the other neighbouring
cell in a hexmatrix.}
\usage{
whichDir(current, other, m)
}
\arguments{
\item{current}{The index of the actual cell in hexmatrix}

\item{other}{The index of the neighbouring cell}

\item{m}{The underlying hexmatrix or hexarray}
}
\value{
A number from 1 to 6 indicating the direction from `current` to
  `other` cell. `NA` is returned if `current` and `other` are not neighbours.
}
\description{
Return the number of direction from the current to the other neighbouring
cell in a hexmatrix.
}
